requirements.txt
aiosmtplib==2.0.2
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
Babel==2.14.0
bcrypt==4.0.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
coverage==7.4.4
cryptography==42.0.5
dnspython==2.6.1
docutils==0.20.1
ecdsa==0.18.0
email_validator==2.1.1
Faker==24.4.0
fastapi==0.110.0
fastapi-limiter==0.1.6
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.3
libgravatar==1.0.4
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pypng==0.20220715.0
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
qrcode==7.4.2
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
